Refactor : 내 장부 조회 #14
서비스와 서비스 사이의 의존성 주입 제거
